Crate media

Modules

media_thread 🔒

Structs

GLPlayerExternalImages 🔒

Bridge between the webrender::ExternalImage callbacks and the GLPlayerThreads.

WindowGLContext

A PlayerGLContext that renders to a window. Note that if the background thread is not started for this context, then it is inactive (returning Unknown values in the trait implementation).

Enums

GLPlayerMsg

GLPlayer thread Message API

GLPlayerMsgForward

These are the messages that the GLPlayer thread will forward to the video player which lives in htmlmediaelement

GlApi

GlContext

NativeDisplay

Statics

WINDOW_GL_CONTEXT 🔒

A global version of the WindowGLContext to be shared between the embedder and the constellation. This is only okay to do because OpenGL contexts cannot be used across processes anyway.

Traits

PlayerGLContext