Crate media

Modules

media_channel
Enum wrappers to be able to select different channel implementations at runtime.

media_thread

Structs

GLPlayerExternalImages
Bridge between the webrender::ExternalImage callbacks and the GLPlayerThreads.

GLPlayerThreads
GLPlayer Threading API entry point that lives in the constellation.

WindowGLContext

Enums

GLPlayerMsg
GLPlayer thread Message API

GLPlayerMsgForward
These are the messages that the GLPlayer thread will forward to the video player which lives in htmlmediaelement

Functions

glplayer_channel